Added configuration option for primary/backup server role
Now that SignatureCascadeStore runs WanAgent, it won't work for the
backup server to be a separate binary that runs WanAgent in the main
thread. Both primary and backup will run SignatureCascadeStore, which
means they will both run the same WanAgent code, so there needs to be a
configuration option that tells SignatureCascadeStore whether it behaves
as the primary or the backup.

Also removed the obsolete wan_backup_udl.

+++ include/cascade/object.hpp
@@ -212,7 +212,7 @@ inline std::ostream& operator<<(std::ostream& out, const Blob& b) {
     if(b.size > 0) {
         uint32_t i = 0;
         for(i = 0; i < 8 && i < b.size; i++) {
-            out << " " << b.bytes[i];
+            out << " " << static_cast<int>(b.bytes[i]);
         }
         if(i < b.size) {
             out << "...";
